fix: Add and update icons, types, and fix preloader drawing
This commit introduces a new 'Pin' icon to the Dashboard component. It also addresses type inconsistencies by making the 'observaciones' property required in the Produccion module and adding the 'creationDate' property to the Ticket interface in ReportesAnaliticas. Additionally, it resolves a drawing error in the Preloader component by correctly specifying the fill rule for the canvas arc.